- ZNTexture 추가 - DirectXTex.lib 추가하여 텍스쳐 로드 ㄴCreateTexture에서 DirectXTex 의 CreateTexture, PrepareTexture 사용 시 링크에러가 해결이 안돼서 CreateCommittedResource로 처리. 프로젝트 속성 및 여러가지 시도를 해봤는데 해결이 안됨. - IndexBuffer 잘못 처리하던 부분 있어서 수정.

+++ ZN/ZNFramework/Window/ZNWindow.h
@@ -14,12 +14,12 @@ namespace ZNFramework
 		virtual void Show() = 0;
 		virtual void Hide() = 0;
 
-		virtual uint32_t Width() const = 0;
-		virtual uint32_t Height() const = 0;
+		virtual uint32 Width() const = 0;
+		virtual uint32 Height() const = 0;
 		virtual void* PlatformHandle() const = 0;
 
 		using EventHandler = const void*;
-		using ResizeEventCallback = std::function<void(size_t, size_t)>;
+		using ResizeEventCallback = std::function<void(uint32, uint32)>;
 
 		void AddEventHandler(EventHandler handler, ResizeEventCallback callback)
 		{
